Fix #149 GUI scrollbar working differently

+++ src/main/java/binnie/core/craftgui/controls/scroll/ControlScrollBar.java
@@ -25,7 +25,7 @@ public class ControlScrollBar extends Control {
 		this.addSelfEventHandler(new EventMouse.Drag.Handler() {
 			@Override
 			public void onEvent(final EventMouse.Drag event) {
-				ControlScrollBar.this.scrollable.movePercentage(event.getDy() / (ControlScrollBar.this.height() - ControlScrollBar.this.getBarHeight()));
+				ControlScrollBar.this.scrollable.movePercentage(event.getDy() / (float) (ControlScrollBar.this.height() - ControlScrollBar.this.getBarHeight()));
 			}
 		});
 		this.addSelfEventHandler(new EventMouse.Down.Handler() {
@@ -35,13 +35,13 @@ public class ControlScrollBar extends Control {
 				final float percentageIndex = ControlScrollBar.this.scrollable.getPercentageIndex();
 				final float minPercent = (1.0f - shownPercentage) * percentageIndex;
 				final float maxPercent = minPercent + shownPercentage;
-				float clickedPercentage = ControlScrollBar.this.getRelativeMousePosition().y() / (ControlScrollBar.this.height() - 2.0f);
+				float clickedPercentage = (float)ControlScrollBar.this.getRelativeMousePosition().y() / (float) (ControlScrollBar.this.height() - 2.0f);
 				clickedPercentage = Math.max(Math.min(clickedPercentage, 1.0f), 0.0f);
 				if (clickedPercentage > maxPercent) {
-					ControlScrollBar.this.scrollable.setPercentageIndex((clickedPercentage - shownPercentage) / (1.0f - shownPercentage));
+					ControlScrollBar.this.scrollable.setPercentageIndex((float) (clickedPercentage - shownPercentage) / (float) (1.0f - shownPercentage));
 				}
 				if (clickedPercentage < minPercent) {
-					ControlScrollBar.this.scrollable.setPercentageIndex(clickedPercentage / (1.0f - shownPercentage));
+					ControlScrollBar.this.scrollable.setPercentageIndex(clickedPercentage / (float) (1.0f - shownPercentage));
 				}
 			}
 		});
